fix: 모델 구현 요구사항에 맞게 모듈 재분리
