+++ src/BridgeGame.js
@@ -1,118 +1,57 @@
 const GAME_SIGNATURE = require('./utils/constant');
 
-/**
- * 다리 건너기 게임을 관리하는 클래스
- */
 class BridgeGame {
-  constructor(
-    { bridgeRandomNumberGenerator, bridgeMaker, bridgeModel },
-    { outputView, inputView, validator }
-  ) {
-    this.bridgeModel = bridgeModel;
-    this.bridgeRandomNumberGenerator = bridgeRandomNumberGenerator;
-    this.bridgeMaker = bridgeMaker;
-    this.outputView = outputView;
-    this.inputView = inputView;
-    this.validator = validator;
-    this.start();
-    this.askBridgeSize();
+  bridge;
+  constructor() {
+    this.trialCount = 1;
+    this.trials = [];
+    this.status = GAME_SIGNATURE.gameOn;
   }
 
-  start() {
-    this.outputView.printStartMessage();
+  setBridge(bridge) {
+    this.bridge = bridge;
   }
 
-  askBridgeSize() {
-    this.inputView.readBridgeSize(this.makeBridge.bind(this));
-  }
-
-  makeBridge(size) {
-    try {
-      this.validator.checkBridgeSize(size);
-    } catch (error) {
-      this.outputView.printError(error.message);
-      return;
-    }
-
-    const bridge = this.bridgeMaker.makeBridge(size, this.bridgeRandomNumberGenerator.generate);
-    this.bridgeModel.setBridge(bridge);
-
-    this.outputView.newLine();
-    this.askMoveDirection();
-  }
-
-  askMoveDirection() {
-    this.inputView.readMoving(this.move.bind(this));
-  }
-
-  /**
-   * 사용자가 칸을 이동할 때 사용하는 메서드
-   * <p>
-   * 이동을 위해 필요한 메서드의 반환 값(return value), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-   */
   move(direction) {
-    try {
-      this.validator.checkMove(direction);
-    } catch (error) {
-      this.outputView.printError(error.message);
-      return;
-    }
-
-    this.bridgeModel.move(direction);
-    this.outputView.printMap(this.bridgeModel.trials);
-
-    if (this.bridgeModel.status === GAME_SIGNATURE.gameOn) {
-      this.askMoveDirection();
-      return;
-    }
-
-    if (this.bridgeModel.status === GAME_SIGNATURE.gameFail) {
-      this.gameCommand();
-      return;
-    }
-
-    if (this.bridgeModel.status === GAME_SIGNATURE.gameSuccess) {
-      this.end();
-      return;
-    }
+    this.updateTrialList(direction, this.getStage());
+    this.checkStatus();
   }
 
-  end() {
-    this.outputView.printResult(this.bridgeModel);
-  }
+  checkStatus() {
+    const lastTrial = [...this.trials].pop();
 
-  gameCommand() {
-    this.inputView.readGameCommand(this.handleCommand.bind(this));
-  }
-
-  handleCommand(command) {
-    try {
-      this.validator.checkGameCommand(command);
-    } catch (error) {
-      this.outputView.printError(error.message);
+    if (lastTrial.result === GAME_SIGNATURE.pass && this.trials.length === this.bridge.length) {
+      this.status = GAME_SIGNATURE.gameSuccess;
       return;
     }
 
-    if (command === 'R') {
-      this.retry();
-
+    if (lastTrial.result === GAME_SIGNATURE.fail) {
+      this.status = GAME_SIGNATURE.gameFail;
       return;
     }
 
-    if (command === 'Q') {
-      this.end();
-    }
+    this.status = GAME_SIGNATURE.gameOn;
   }
 
-  /**
-   * 사용자가 게임을 다시 시도할 때 사용하는 메서드
-   * <p>
-   * 재시작을 위해 필요한 메서드의 반환 값(return value), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-   */
+  getStage() {
+    return this.trials.length;
+  }
+
+  getTrialResult(trialDirection, stage) {
+    const answerDirection = this.bridge[stage];
+
+    return answerDirection === trialDirection ? GAME_SIGNATURE.pass : GAME_SIGNATURE.fail;
+  }
+
+  updateTrialList(trialDirection, stage) {
+    this.trials.push({
+      direction: trialDirection,
+      result: this.getTrialResult(trialDirection, stage),
+    });
+  }
   retry() {
-    this.bridgeModel.trialCount += 1;
-    this.bridgeModel.trials = [];
-    this.askMoveDirection();
+    this.trialCount += 1;
+    this.trials = [];
   }
 }
