+++ src/views/OutputView.js
@@ -54,10 +54,10 @@ const OutputView = {
    * <p>
    * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
    */
-  printResult(bridgeModel) {
-    const [bridgeUpperSide, bridgeLowerSide] = this.getBridgeMap(bridgeModel.trials);
-    const trialCount = bridgeModel.trialCount;
-    const status = bridgeModel.status;
+  printResult(bridgeGame) {
+    const [bridgeUpperSide, bridgeLowerSide] = this.getBridgeMap(bridgeGame.trials);
+    const trialCount = bridgeGame.trialCount;
+    const status = bridgeGame.status;
     Console.print(
       `최종 게임 결과
 ${bridgeUpperSide}
